+++ src/main/java/com/pagely/paymentservice/infrastructure/client/pg/dto/TossConfirmRequest.java
@@ -1,4 +1,4 @@
-package com.pagely.paymentservice.infrastructure.client.pg;
+package com.pagely.paymentservice.infrastructure.client.pg.dto;
 
 public record TossConfirmRequest(
         String paymentKey,
